array: move out index operation into separate file
